+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Fix passing absolute .wav paths into corrscope CLI (#398)
 - Fix preview error when clearing "Trigger/Render Width" table cells (#407)
+- Add control for DC removal rate (#408)
 
 ## 0.7.1